GHForkDemo/Controllers/HomeController.cs
using GHForkDemo.Models;
using GHForkDemo.Shared.Models.Home;
using Microsoft.AspNetCore.Mvc;
using System.Diagnostics;
namespace GHForkDemo.Controllers
{
    public class HomeController : Controller
    {
        private readonly ILogger<HomeController> _logger;

        public HomeController(ILogger<HomeController> logger)
        {
            _logger = logger;
        }

        public IActionResult Index(string name = "Index")
        {
            return View(new IndexModel
            {
                Name = name
            });
        }

        public IActionResult Privacy()
        {
            return View();
        }

        [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
        public IActionResult Error()
        {
            return View(new ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });
        }
    }
}